# packages/regpoly-web/tests/fixtures/openapi.v2.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/generators": {
      "get": {
        "operationId": "v2_list_generators_api_v2_generators_get",
        "parameters": [
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "in": "query",
            "name": "k",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "K"
            }
          },
          {
            "in": "query",
            "name": "k_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "K Min"
            }
          },
          {
            "in": "query",
            "name": "k_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "K Max"
            }
          },
          {
            "in": "query",
            "name": "search_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Run Id"
            }
          },
          {
            "in": "query",
            "name": "has_search_results",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Search Results"
            }
          },
          {
            "description": "primitive yes|no",
            "in": "query",
            "name": "primitive",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(yes|no)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "primitive yes|no",
              "title": "Primitive"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratorList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 List Generators",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/api/v2/import/generators": {
      "post": {
        "description": "Accept a multipart upload, persist to a tmp file, and route through\nthe existing v1 single-file importer. Replaces the broken JS-only\n`commitImport()` flow on /tools.",
        "operationId": "v2_import_generators_api_v2_import_generators_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_v2_import_generators_api_v2_import_generators_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Import Generators",
        "tags": [
          "v2"
        ]
      }
    },
    "/api/v2/import/generators/preview": {
      "post": {
        "operationId": "v2_import_preview_api_v2_import_generators_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_v2_import_preview_api_v2_import_generators_preview_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Import Preview",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/api/v2/tested-generators": {
      "get": {
        "operationId": "v2_list_tested_generators_api_v2_tested_generators_get",
        "parameters": [
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "in": "query",
            "name": "k_g_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "K G Min"
            }
          },
          {
            "in": "query",
            "name": "k_g_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "K G Max"
            }
          },
          {
            "in": "query",
            "name": "test_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Test Type"
            }
          },
          {
            "in": "query",
            "name": "max_sum_delta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Sum Delta"
            }
          },
          {
            "in": "query",
            "name": "is_me",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Me"
            }
          },
          {
            "in": "query",
            "name": "has_results",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Results"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestedGeneratorList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 List Tested Generators",
        "tags": [
          "v2"
        ]
      }
    },
...
    "/api/v2/tested-generators/{tg_id}": {
      "get": {
        "description": "Full-detail companion to the v2 list/publish endpoints. Returns\nthe row + components + per-test results in one call so an SDK\nconsumer doesn't need to fan out across v1+v2 namespaces.",
        "operationId": "v2_get_tested_generator_api_v2_tested_generators__tg_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tg_id",
            "required": true,
            "schema": {
              "title": "Tg Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestedGeneratorDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Get Tested Generator",
        "tags": [
          "v2"
        ]
      }
    },
...
  }
}